feat(test): use connection strings for all calls to `newClient`
We have been inconsistently testing our code for some time using
a deprecated form of creating a `MongoClient` by passing in a
topology to the constructor. Now we take the passed in variables
and construct a connection string to connect to the server, just
like users of the module would.